requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.12.1
click==8.1.7
Django==4.2.8
djangorestframework==3.14.0
exceptiongroup==1.2.0
flake8==7.0.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==7.4.4
pytest-django==4.7.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
